requirements.txt
django
gunicorn
django-heroku
psycopg2
djangorestframework
markdown
django-filter
django-cors-headers
nltk
sumy
